+++ internal/cortex/querier/queryrange/test_utils_test.go
@@ -9,6 +9,7 @@ import (
 	"testing"
 
 	"github.com/prometheus/prometheus/model/labels"
+	"github.com/prometheus/prometheus/tsdb/chunkenc"
 	"github.com/stretchr/testify/require"
 
 	"github.com/thanos-io/thanos/internal/cortex/querier/astmapper"
@@ -126,7 +127,7 @@ func TestNewMockShardedqueryable(t *testing.T) {
 				seriesCt++
 				iter := set.At().Iterator()
 				samples := 0
-				for iter.Next() {
+				for iter.Next() != chunkenc.ValNone {
 					samples++
 				}
 				require.Equal(t, tc.nSamples, samples)