+++ pkg/dedup/iter_test.go
@@ -11,11 +11,12 @@ import (
 	"sort"
 	"testing"
 
+	"github.com/prometheus/prometheus/model/histogram"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/storage"
 	"github.com/prometheus/prometheus/tsdb/chunkenc"
 
-	"github.com/thanos-io/thanos/pkg/testutil"
+	"github.com/efficientgo/core/testutil"
 )
 
 type sample struct {
@@ -29,6 +30,19 @@ func (s sample) T() int64 {
 
 func (s sample) V() float64 {
 	return s.v
+}
+
+// TODO(rabenhorst): Needs to be implemented for native histogram support.
+func (s sample) H() *histogram.Histogram {
+	panic("not implemented")
+}
+
+func (s sample) FH() *histogram.FloatHistogram {
+	panic("not implemented")
+}
+
+func (s sample) Type() chunkenc.ValueType {
+	return chunkenc.ValFloat
 }
 
 type series struct {
@@ -68,11 +82,15 @@ func newMockedSeriesIterator(samples []sample) *mockedSeriesIterator {
 	return &mockedSeriesIterator{samples: samples, cur: -1}
 }
 
-func (s *mockedSeriesIterator) Seek(t int64) bool {
+// TODO(rabenhorst): Native histogram support needs to be added, currently hardcoded to float.
+func (s *mockedSeriesIterator) Seek(t int64) chunkenc.ValueType {
 	s.cur = sort.Search(len(s.samples), func(n int) bool {
 		return s.samples[n].t >= t
 	})
-	return s.cur < len(s.samples)
+	if s.cur < len(s.samples) {
+		return chunkenc.ValFloat
+	}
+	return chunkenc.ValNone
 }
 
 func (s *mockedSeriesIterator) At() (t int64, v float64) {
@@ -80,9 +98,26 @@ func (s *mockedSeriesIterator) At() (t int64, v float64) {
 	return sample.t, sample.v
 }
 
-func (s *mockedSeriesIterator) Next() bool {
+// TODO(rabenhorst): Needs to be implemented for native histogram support.
+func (s *mockedSeriesIterator) AtHistogram() (int64, *histogram.Histogram) {
+	panic("not implemented")
+}
+
+func (s *mockedSeriesIterator) AtFloatHistogram() (int64, *histogram.FloatHistogram) {
+	panic("not implemented")
+}
+
+func (s *mockedSeriesIterator) AtT() int64 {
+	return s.samples[s.cur].t
+}
+
+func (s *mockedSeriesIterator) Next() chunkenc.ValueType {
 	s.cur++
-	return s.cur < len(s.samples)
+	if s.cur < len(s.samples) {
+		return chunkenc.ValFloat
+	}
+
+	return chunkenc.ValNone
 }
 
 func (s *mockedSeriesIterator) Err() error { return nil }
@@ -497,7 +532,7 @@ func BenchmarkDedupSeriesIterator(b *testing.B) {
 		b.ResetTimer()
 		var total int64
 
-		for it.Next() {
+		for it.Next() != chunkenc.ValNone {
 			t, _ := it.At()
 			total += t
 		}
@@ -541,7 +576,7 @@ func BenchmarkDedupSeriesIterator(b *testing.B) {
 const hackyStaleMarker = float64(-99999999)
 
 func expandSeries(t testing.TB, it chunkenc.Iterator) (res []sample) {
-	for it.Next() {
+	for it.Next() != chunkenc.ValNone {
 		t, v := it.At()
 		// Nan != Nan, so substitute for another value.
 		// This is required for testutil.Equals to work deterministically.