+++ pkg/store/cache/inmemory_test.go
@@ -19,7 +19,7 @@ import (
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/storage"
 
-	"github.com/thanos-io/thanos/pkg/testutil"
+	"github.com/efficientgo/core/testutil"
 )
 
 func TestNewInMemoryIndexCache(t *testing.T) {