+++ test/e2e/info_api_test.go
@@ -18,9 +18,9 @@ import (
 
 	"github.com/thanos-io/objstore/client"
 
+	"github.com/efficientgo/core/testutil"
 	"github.com/thanos-io/thanos/pkg/query"
 	"github.com/thanos-io/thanos/pkg/runutil"
-	"github.com/thanos-io/thanos/pkg/testutil"
 	"github.com/thanos-io/thanos/test/e2e/e2ethanos"
 )